backend/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
astroid==2.14.2
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-excel==0.0.10
django-rest-multiple-models==2.1.3
django-simple-history==3.2.0
django-storages==1.13.2
django-types==0.17.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-core==1.34.0
google-auth==2.16.3
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphene==3.2.2
graphene-django==3.0.0b7
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
lml==0.1.0
mccabe==0.7.0
pathspec==0.11.1
platformdirs==3.0.0
promise==2.3
protobuf==3.20.3
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.6.0
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
texttable==1.6.7
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
